Highlighting colors

+++ lib/views/mbx_relogin_screen/mbx_relogin_controller.dart
@@ -45,13 +45,15 @@ class MbxReloginController extends GetxController {
             return false;
           }
         }).then((code) {
-      LoggerX.log('[PIN] verfied: $code');
-      Get.loading();
-      MbxProfileVM.request().then((resp) {
-        Get.back();
-        Get.deleteAll();
-        Get.offAll(MbxBottomNavBarScreen());
-      });
+      if (code != null && (code as String).isNotEmpty) {
+        LoggerX.log('[PIN] verfied: $code');
+        Get.loading();
+        MbxProfileVM.request().then((resp) {
+          Get.back();
+          Get.deleteAll();
+          Get.offAll(MbxBottomNavBarScreen());
+        });
+      }
     });
   }
